move to more central file detector
