+++ src/Microsoft.ComponentDetection.Detectors/pip/PipReportComponentDetector.cs
@@ -23,12 +23,8 @@ public class PipReportComponentDetector : FileComponentDetector
     private const string PipReportSkipFallbackOnFailureEnvVar = "PipReportSkipFallbackOnFailure";
     private const string PipReportFileLevelTimeoutSecondsEnvVar = "PipReportFileLevelTimeoutSeconds";
     private const string PipReportPersistReportsEnvVar = "PipReportPersistReports";
-    private const string PipReportCleanCreatedFilesDryRunEnvVar = "PipReportCleanCreatedFilesDryRun";
-    private const int DefaultCleanDepth = 5;
 
     private static readonly IList<string> PipReportPreGeneratedFilePatterns = new List<string> { "*.component-detection-pip-report.json", "component-detection-pip-report.json" };
-
-    private static readonly IList<string> CleanUpPatterns = new List<string> { "*.egg", "*.egg-info", "*.pyc", "*.pyo", "*.pyd", "__pycache__" };
 
     /// <summary>
     /// The maximum version of the report specification that this detector can handle.
@@ -79,11 +75,13 @@ public class PipReportComponentDetector : FileComponentDetector
 
     public override IEnumerable<string> Categories => new List<string> { "Python" };
 
-    public override IEnumerable<ComponentType> SupportedComponentTypes { get; } = new[] { ComponentType.Pip };
+    public override IEnumerable<ComponentType> SupportedComponentTypes { get; } = new List<ComponentType> { ComponentType.Pip };
 
     public override int Version { get; } = 8;
 
     protected override bool EnableParallelism { get; set; } = true;
+
+    protected override IList<string> CleanUpPatterns => new List<string> { "*.egg", "*.egg-info", "*.pyc", "*.pyo", "*.pyd", "__pycache__" };
 
     protected override async Task<IObservable<ProcessRequest>> OnPrepareDetectionAsync(IObservable<ProcessRequest> processRequests, IDictionary<string, string> detectorArgs, CancellationToken cancellationToken = default)
     {
@@ -131,10 +129,7 @@ public class PipReportComponentDetector : FileComponentDetector
         var singleFileComponentRecorder = processRequest.SingleFileComponentRecorder;
         var file = processRequest.ComponentStream;
 
-        DirectoryInfo fileParentDirectoryInfo = null;
         List<FileInfo> reportFiles = new();
-        HashSet<string> preExistingFiles = new();
-        HashSet<string> preExistingDirs = new();
         try
         {
             var pipOverride = this.GetPipReportOverrideBehavior();
@@ -180,15 +175,13 @@ public class PipReportComponentDetector : FileComponentDetector
                 return;
             }
 
-            fileParentDirectoryInfo = Directory.Exists(fileParentDirectory)
+            var fileParentDirectoryInfo = Directory.Exists(fileParentDirectory)
                 ? new DirectoryInfo(fileParentDirectory)
                 : null;
 
             List<FileInfo> preGeneratedReportFiles = null;
             if (fileParentDirectoryInfo is not null)
             {
-                (preExistingFiles, preExistingDirs) = this.fileUtilityService.GetFilesAndDirectories(fileParentDirectory, CleanUpPatterns, DefaultCleanDepth);
-
                 preGeneratedReportFiles = PipReportPreGeneratedFilePatterns
                     .SelectMany(pattern => fileParentDirectoryInfo.GetFiles(pattern))
                     .Where(file => File.Exists(file.FullName))
@@ -327,55 +320,6 @@ public class PipReportComponentDetector : FileComponentDetector
                     if (reportFile is not null && reportFile.Exists)
                     {
                         reportFile.Delete();
-                    }
-                }
-            }
-
-            // Clean up any new files or directories created during the scan that match the clean up patters
-            if (fileParentDirectoryInfo is not null)
-            {
-                var dryRun = this.envVarService.IsEnvironmentVariableValueTrue(PipReportCleanCreatedFilesDryRunEnvVar);
-                var dryRunStr = dryRun ? "[DRYRUN]" : string.Empty;
-                var (newFiles, newDirs) = this.fileUtilityService.GetFilesAndDirectories(fileParentDirectoryInfo.FullName, CleanUpPatterns, DefaultCleanDepth);
-                var createdFiles = newFiles.Except(preExistingFiles).ToList();
-                var createdDirs = newDirs.Except(preExistingDirs).ToList();
-
-                foreach (var createdDir in createdDirs)
-                {
-                    if (createdDir is null || !Directory.Exists(createdDir))
-                    {
-                        continue;
-                    }
-
-                    try
-                    {
-                        this.Logger.LogDebug("{DryRun} PipReport: Cleaning up directory {Dir}", dryRunStr, createdDir);
-                        if (!dryRun)
-                        {
-                            Directory.Delete(createdDir, true);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        this.Logger.LogDebug(e, "{DryRun} PipReport: Failed to delete directory {Dir}", dryRunStr, createdDir);
-                    }
-                }
-
-                foreach (var createdFile in createdFiles)
-                {
-                    if (createdFile is null || !File.Exists(createdFile))
-                    {
-                        continue;
-                    }
-
-                    try
-                    {
-                        this.Logger.LogDebug("{DryRun} PipReport: Cleaning up file {File}", dryRunStr, createdFile);
-                        File.Delete(createdFile);
-                    }
-                    catch (Exception e)
-                    {
-                        this.Logger.LogDebug(e, "{DryRun} PipReport: Failed to delete file {File}", dryRunStr, createdFile);
                     }
                 }
             }
